gallery/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
jinja2>=3.1.0
google-cloud-storage>=2.0.0
google-auth>=2.0.0
Pillow>=10.0.0
